Extended Twig to receive user defined functions
Added the possibility to extend Twig with user defined functions without modifying Twig.php directly.

The optional 4th parameter consists of an array of Twig_SimpleFunction elements to be merged with the default ones.
Duplicity is possible and not forbidden, this allows overriding default functionality as the last defined element will take priority over the other(s).